requirements.txt
Django>=4.2.0,<5.0
Pillow>=10.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
python-decouple>=3.8
opencv-python>=4.8.0
numpy>=1.24.0 